Refactoring nodes in its own module
